+++ apps/dsp-app/cypress.config.ts
@@ -2,9 +2,13 @@
 import { defineConfig } from 'cypress';
 
 export default defineConfig({
+  projectId: 'n5b5id',
   e2e: {
     specPattern: 'cypress/**/**/**/*.cy.ts',
     excludeSpecPattern: ['*.spec.js', '*.spec.ts'],
+    viewportHeight: 768,
+    viewportWidth: 1024,
+    supportFile: 'cypress/support/e2e.ts',
     baseUrl: 'http://localhost:4200',
     env: {
       apiUrl: 'http://0.0.0.0:3333',