+++ apps/dsp-app/src/app/project/image-settings/image-settings.component.ts
@@ -1,7 +1,6 @@
 import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { ProjectRestrictedViewSettings } from '@dasch-swiss/dsp-js';
-import { SetRestrictedViewRequest } from '@dasch-swiss/vre/open-api';
 import { RouteConstants } from '@dasch-swiss/vre/shared/app-config';
 import { ProjectService } from '@dasch-swiss/vre/shared/app-helper-services';
 import { NotificationService } from '@dasch-swiss/vre/shared/app-notification';
@@ -13,8 +12,15 @@ import {
 import { TranslateService } from '@ngx-translate/core';
 import { Select, Store } from '@ngxs/store';
 import { Observable } from 'rxjs';
-import { switchMap, takeWhile } from 'rxjs/operators';
+import { switchMap, take, takeWhile } from 'rxjs/operators';
 import { ReplaceAnimation } from '../../main/animations/replace-animation';
+import { InputMasks } from '../../main/directive/input-masks';
+
+enum ImageSettingsEnum {
+  Off = 'Off',
+  Watermark = 'Watermark',
+  RestrictImageSize = 'RestrictImageSize',
+}
 
 @Component({
   selector: 'app-image-settings',
@@ -23,18 +29,35 @@ import { ReplaceAnimation } from '../../main/animations/replace-animation';
   animations: [ReplaceAnimation.animation],
 })
 export class ImageSettingsComponent implements OnInit {
-  readonly absoluteWidthSteps: number[] = [64, 128, 256, 512, 1024];
+  readonly minWidth = 128;
+  readonly maxWidth = 1024;
+  readonly inputMasks = InputMasks;
+  readonly imageSettingsEnum = ImageSettingsEnum;
+
+  currentSettings: ProjectRestrictedViewSettings;
+  imageSettings: ImageSettingsEnum = ImageSettingsEnum.Off;
   projectUuid = this.route.parent.parent.snapshot.paramMap.get(RouteConstants.uuidParameter);
+  percentage: string = '99';
+  fixedWidth: string;
+
+  get hasChanges(): boolean {
+    return JSON.stringify(this.currentSettings) !== JSON.stringify(this.getRequest());
+  }
+
+  get ratio(): number {
+    if (!this.percentage) {
+      return 0;
+    }
+
+    return parseInt(this.percentage, 0) / 100;
+  }
+
+  get isPercentageSize(): boolean {
+    return this.percentage !== null;
+  }
 
   @Select(ProjectsSelectors.isProjectsLoading) isProjectsLoading$: Observable<boolean>;
-  @Select(ProjectsSelectors.projectRestrictedViewSettings)
-  projectRestrictedViewSettings$: Observable<ProjectRestrictedViewSettings>;
-
-  percentage = 80;
-  fixedWidth = this.absoluteWidthSteps[3];
-  isWatermark = false;
-  allowRestriction = false;
-  isPercentageSize = true;
+  @Select(ProjectsSelectors.projectRestrictedViewSettings) viewSettings$: Observable<ProjectRestrictedViewSettings>;
 
   constructor(
     private _projectService: ProjectService,
@@ -46,55 +69,84 @@ export class ImageSettingsComponent implements OnInit {
   ) {}
 
   ngOnInit() {
+    this.getImageSettings();
+  }
+
+  onSubmit() {
+    this._store
+      .dispatch(new UpdateProjectRestrictedViewSettingsAction(this.projectUuid, this.getRequest()))
+      .subscribe(() => {
+        this.currentSettings = this._store.selectSnapshot(ProjectsSelectors.projectRestrictedViewSettings);
+        this._notification.openSnackBar(
+          this.translateService.instant('appLabels.form.project.imageSettings.updateConfirmation')
+        );
+      });
+  }
+
+  onPercentageInputChange() {
+    this.fixedWidth = null;
+  }
+
+  onFixedWidthInputChange() {
+    this.percentage = null;
+  }
+
+  private getRequest(): any {
+    return this.imageSettings === ImageSettingsEnum.Watermark
+      ? { watermark: true }
+      : { size: this.getSizeForRequest() };
+  }
+
+  private getImageSettings() {
     this._store
       .dispatch(new LoadProjectRestrictedViewSettingsAction(this._projectService.uuidToIri(this.projectUuid)))
       .pipe(
-        switchMap(() => this.projectRestrictedViewSettings$.pipe(takeWhile(settings => settings?.watermark !== null)))
+        switchMap(() =>
+          this.viewSettings$.pipe(
+            take(1),
+            takeWhile(settings => settings !== null)
+          )
+        )
       )
       .subscribe(settings => {
-        this.isWatermark = settings.watermark as boolean;
-
-        if (settings.size === 'pct:100') {
-          this.allowRestriction = false;
+        if (!settings.watermark && settings.size) {
+          delete settings.watermark;
+        } else if (!settings.watermark) {
           return;
         }
 
-        this.allowRestriction = true;
-        this.isPercentageSize = settings.size.startsWith('pct');
-        if (this.isPercentageSize) {
-          this.percentage = parseInt(settings.size.split(':')[1], 0);
-        } else {
-          this.fixedWidth = parseInt(settings.size.split(',')[1], 0);
+        this.currentSettings = settings;
+        if ((settings.watermark === false && !settings.size) || (settings.size && settings.size === 'pct:100')) {
+          this.imageSettings = this.imageSettingsEnum.Off;
+          return;
         }
+
+        if (settings.watermark) {
+          this.imageSettings = ImageSettingsEnum.Watermark;
+        } else {
+          this.setRestrictedSize(settings.size);
+        }
+
         this._cd.detectChanges();
       });
   }
 
-  formatPercentageLabel = (value: number): string => `${value}%`;
+  private setRestrictedSize(size: string) {
+    if (size) {
+      this.imageSettings = ImageSettingsEnum.RestrictImageSize;
+    }
 
-  formatAbsoluteLabel = (value: number): string => `${this.absoluteWidthSteps[value]}px`;
-
-  absoluteWidthIndex = (value: number): number => this.absoluteWidthSteps.findIndex(step => step == value);
-
-  absoluteSliderChange(value: number) {
-    this.fixedWidth = this.absoluteWidthSteps[value];
-  }
-
-  onSubmit() {
-    const request: SetRestrictedViewRequest = {
-      size: this.getSizeForRequest(),
-      watermark: this.isWatermark,
-    };
-
-    this._store.dispatch(new UpdateProjectRestrictedViewSettingsAction(this.projectUuid, request)).subscribe(() => {
-      this._notification.openSnackBar(
-        this.translateService.instant('appLabels.form.project.imageSettings.updateConfirmation')
-      );
-    });
+    if (size.startsWith('pct')) {
+      this.percentage = size.split(':')[1];
+      this.fixedWidth = null;
+    } else {
+      this.fixedWidth = size.split(',')[1];
+      this.percentage = null;
+    }
   }
 
   private getSizeForRequest() {
-    if (this.allowRestriction) {
+    if (this.imageSettings === ImageSettingsEnum.RestrictImageSize) {
       return this.isPercentageSize ? `pct:${this.percentage}` : `!${this.fixedWidth},${this.fixedWidth}`;
     } else {
       return 'pct:100';