+++ apps/dsp-app/src/app/project/ontology/resource-class-info/resource-class-property-info/resource-class-property-info.component.ts
@@ -1,4 +1,14 @@
-import { AfterContentInit, Component, EventEmitter, Inject, Input, OnChanges, Output } from '@angular/core';
+import {
+  AfterContentInit,
+  ChangeDetectorRef,
+  Component,
+  EventEmitter,
+  Inject,
+  Input,
+  OnChanges,
+  OnDestroy,
+  Output,
+} from '@angular/core';
 import {
   CanDoResponse,
   Constants,
@@ -12,6 +22,8 @@ import { DspApiConnectionToken } from '@dasch-swiss/vre/shared/app-config';
 import { DefaultClass, DefaultProperty, OntologyService } from '@dasch-swiss/vre/shared/app-helper-services';
 import { ListsSelectors, OntologiesSelectors } from '@dasch-swiss/vre/shared/app-state';
 import { Store } from '@ngxs/store';
+import { Subject, Subscription } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 // property data structure
 export class Property {
@@ -44,7 +56,7 @@ type CardinalityKey = 'multiple' | 'required';
   templateUrl: './resource-class-property-info.component.html',
   styleUrls: ['./resource-class-property-info.component.scss'],
 })
-export class ResourceClassPropertyInfoComponent implements OnChanges, AfterContentInit {
+export class ResourceClassPropertyInfoComponent implements OnChanges, AfterContentInit, OnDestroy {
   @Input() propDef: ResourcePropertyDefinitionWithAllLanguages;
 
   @Input() propCard: IHasProperty;
@@ -69,6 +81,8 @@ export class ResourceClassPropertyInfoComponent implements OnChanges, AfterConte
     targetCardinality: GuiCardinality;
   }>();
 
+  isDestroyed = new Subject<void>();
+
   propInfo: Property = new Property();
 
   propType: DefaultProperty;
@@ -77,12 +91,12 @@ export class ResourceClassPropertyInfoComponent implements OnChanges, AfterConte
   propAttributeComment: string;
 
   propCanBeRemovedFromClass: boolean;
-
   constructor(
     @Inject(DspApiConnectionToken)
     private _dspApiConnection: KnoraApiConnection,
     private _ontoService: OntologyService,
-    private _store: Store
+    private _store: Store,
+    private _cd: ChangeDetectorRef
   ) {}
 
   ngOnChanges(): void {
@@ -133,16 +147,22 @@ export class ResourceClassPropertyInfoComponent implements OnChanges, AfterConte
     }
 
     // get current ontology lists to get linked list information
-    const currentOntologyLists = this._store.selectSnapshot(ListsSelectors.listsInProject);
-    if (currentOntologyLists && this.propDef.objectType === Constants.ListValue) {
-      // this property is a list property
-      const re = /\<([^)]+)\>/;
-      const listIri = this.propDef.guiAttributes[0].match(re)[1];
-      const listUrl = `/project/${this.projectUuid}/list/${listIri.split('/').pop()}`;
-      const list = currentOntologyLists.find(i => i.id === listIri);
-      this.propAttribute = `<a href="${listUrl}">${list.labels[0].value}</a>`;
-      this.propAttributeComment = list.comments.length ? list.comments[0].value : null;
-    }
+    this._store
+      .select(ListsSelectors.listsInProject)
+      .pipe(takeUntil(this.isDestroyed))
+      .subscribe(currentOntologyLists => {
+        if (currentOntologyLists && this.propDef.objectType === Constants.ListValue) {
+          // this property is a list property
+          const re = /\<([^)]+)\>/;
+          const listIri = this.propDef.guiAttributes[0].match(re)[1];
+          const listUrl = `/project/${this.projectUuid}/list/${listIri.split('/').pop()}`;
+          const list = currentOntologyLists.find(i => i.id === listIri);
+          if (list) {
+            this.propAttribute = `<a href="${listUrl}">${list.labels[0].value}</a>`;
+            this.propAttributeComment = list.comments.length ? list.comments[0].value : null;
+          }
+        }
+      });
   }
 
   /**
@@ -179,6 +199,12 @@ export class ResourceClassPropertyInfoComponent implements OnChanges, AfterConte
 
     this._dspApiConnection.v2.onto.canDeleteCardinalityFromResourceClass(onto).subscribe((canDoRes: CanDoResponse) => {
       this.propCanBeRemovedFromClass = canDoRes.canDo;
+      this._cd.markForCheck();
     });
+  }
+
+  ngOnDestroy() {
+    this.isDestroyed.next();
+    this.isDestroyed.complete();
   }
 }