+++ apps/dsp-app/src/app/project/reusable-project-form/reusable-project-form.component.ts
@@ -1,12 +1,11 @@
-import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { FormBuilder, Validators } from '@angular/forms';
 import { ProjectsSelectors } from '@dasch-swiss/vre/shared/app-state';
-import { MultiLanguages } from '@dasch-swiss/vre/shared/app-string-literal';
+import { DEFAULT_MULTILANGUAGE_FORM, MultiLanguages } from '@dasch-swiss/vre/shared/app-string-literal';
 import { Store } from '@ngxs/store';
-import { Subscription } from 'rxjs';
-import { startWith } from 'rxjs/operators';
 import { arrayLengthGreaterThanZeroValidator } from '../../main/form-validators/array-length-greater-than-zero-validator';
 import { atLeastOneStringRequired } from '../../main/form-validators/at-least-one-string-required.validator';
+import { ProjectForm } from './project-form.type';
 import { shortcodeExistsValidator } from './shortcode-exists.validator';
 
 @Component({
@@ -15,16 +14,14 @@ import { shortcodeExistsValidator } from './shortcode-exists.validator';
     <form *ngIf="form" [formGroup]="form">
       <div style="display: flex">
         <app-common-input
-          [formGroup]="form"
-          controlName="shortcode"
+          [control]="form.controls.shortcode"
           [placeholder]="'appLabels.form.project.general.shortcode' | translate"
           [validatorErrors]="[shortcodePatternError, shortCodeExistsError]"
           data-cy="shortcode-input"
           style="flex: 1; margin-right: 16px"></app-common-input>
 
         <app-common-input
-          [formGroup]="form"
-          controlName="shortname"
+          [control]="form.controls.shortname"
           [placeholder]="'appLabels.form.project.general.shortname' | translate"
           data-cy="shortname-input"
           style="flex: 1"></app-common-input>
@@ -32,25 +29,25 @@ import { shortcodeExistsValidator } from './shortcode-exists.validator';
 
       <app-common-input
         [placeholder]="'appLabels.form.project.general.longname' | translate"
-        controlName="longname"
-        data-cy="longname-input"
-        [formGroup]="form"></app-common-input>
+        [control]="form.controls.longname"
+        data-cy="longname-input"></app-common-input>
 
       <dasch-swiss-multi-language-textarea
         [placeholder]="('appLabels.form.project.general.description' | translate) + '*'"
-        [formGroup]="form"
-        data-cy="description-input"
-        controlName="description"
-        [validators]="descriptionValidators">
+        [formArray]="form.controls.description"
+        [validators]="descriptionValidators"
+        data-cy="description-input">
       </dasch-swiss-multi-language-textarea>
 
-      <app-chip-list-input [formGroup]="form" controlName="keywords" data-cy="keywords-input" editable="true">
+      <app-chip-list-input
+        [formArray]="form.controls.keywords"
+        data-cy="keywords-input"
         [validators]="keywordsValidators">
       </app-chip-list-input>
     </form>
   `,
 })
-export class ReusableProjectFormComponent implements OnInit, OnDestroy {
+export class ReusableProjectFormComponent implements OnInit {
   @Input() formData: {
     shortcode: string;
     shortname: string;
@@ -58,14 +55,16 @@ export class ReusableProjectFormComponent implements OnInit, OnDestroy {
     description: MultiLanguages;
     keywords: string[];
   };
-  @Output() formValueChange = new EventEmitter<FormGroup>();
+  @Output() afterFormInit = new EventEmitter<ProjectForm>();
 
-  form: FormGroup;
-  shortcodePatternError = { errorKey: 'pattern', message: 'This field must contains letters from A to F and 0 to 9' };
-  shortCodeExistsError = { errorKey: 'shortcodeExists', message: 'This shortcode already exists' };
+  form: ProjectForm;
+  readonly shortcodePatternError = {
+    errorKey: 'pattern',
+    message: 'This field must contains letters from A to F and 0 to 9',
+  };
+  readonly shortCodeExistsError = { errorKey: 'shortcodeExists', message: 'This shortcode already exists' };
   readonly keywordsValidators = [Validators.minLength(3), Validators.maxLength(64)];
   readonly descriptionValidators = [Validators.minLength(3), Validators.maxLength(40960)];
-  subscription: Subscription;
 
   constructor(
     private _fb: FormBuilder,
@@ -74,10 +73,7 @@ export class ReusableProjectFormComponent implements OnInit, OnDestroy {
 
   ngOnInit() {
     this._buildForm();
-
-    this.subscription = this.form.valueChanges.pipe(startWith(<FormGroup>null)).subscribe(() => {
-      this.formValueChange.emit(this.form);
-    });
+    this.afterFormInit.emit(this.form);
   }
 
   private _buildForm() {
@@ -99,15 +95,9 @@ export class ReusableProjectFormComponent implements OnInit, OnDestroy {
         [Validators.required, Validators.minLength(3), Validators.maxLength(20), Validators.pattern(/^[a-zA-Z]+\S*$/)],
       ],
       longname: [this.formData.longname, [Validators.required, Validators.minLength(3), Validators.maxLength(256)]],
-      description: this._fb.array(
-        this.formData.description.map(({ language, value }) =>
-          this._fb.group({
-            language,
-            value: [value, this.descriptionValidators],
-          })
-        ),
-        atLeastOneStringRequired('value')
-      ),
+      description: DEFAULT_MULTILANGUAGE_FORM(this.formData.description, this.descriptionValidators, [
+        atLeastOneStringRequired('value'),
+      ]),
       keywords: this._fb.array(
         this.formData.keywords.map(keyword => {
           return [keyword, this.keywordsValidators];
@@ -115,9 +105,5 @@ export class ReusableProjectFormComponent implements OnInit, OnDestroy {
         arrayLengthGreaterThanZeroValidator()
       ),
     });
-  }
-
-  ngOnDestroy() {
-    this.subscription.unsubscribe();
   }
 }