+++ libs/vre/shared/app-state/src/store.module.ts
@@ -6,9 +6,17 @@ import { NgxsModule } from '@ngxs/store';
 import { CurrentPageState } from './lib/current-page/current-page.state';
 import { CustomRouterStateSerializer } from './lib/router/router-state.serializer';
 import { DEVTOOLS_REDUX_CONFIG, LOGGER_CONFIG, OPTIONS_CONFIG } from './store.config';
-import { ListsState, OntologiesState, OntologyClassState, ProjectsState, UserState } from './index';
+import { ListsState, OntologiesState, OntologyClassState, ProjectsState, ResourceState, UserState } from './index';
 
-const STATE_MODULES = [UserState, ProjectsState, CurrentPageState, OntologiesState, ListsState, OntologyClassState];
+const STATE_MODULES = [
+  UserState,
+  ProjectsState,
+  CurrentPageState,
+  OntologiesState,
+  ListsState,
+  OntologyClassState,
+  ResourceState,
+];
 
 @NgModule({
   exports: [NgxsModule],